15_manejo_mysql/mysql_dev-producion/app/requirements.txt
mysql-connector-python
python-decouple
Flask

